requirements.txt
asn1crypto==0.24.0
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
contextlib2==0.5.5
cryptography==2.3.1
Flask==1.0.2
Flask-OpenTracing==0.2.0
future==0.16.0
google-api-core==1.3.0
google-auth==1.5.1
googleapis-common-protos==1.5.3
idna==2.7
itsdangerous==0.24
jaeger-client==3.11.0
Jinja2==2.10
MarkupSafe==1.0
mysql-connector==2.1.6
mysqlclient==1.3.13
opencensus==0.1.6
opentracing==1.3.0
opentracing-instrumentation==2.4.3
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
requests==2.19.1
rsa==3.4.2
six==1.11.0
threadloop==1.0.2
thrift==0.11.0
tornado==4.5.3
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
